Location added to the details table. PlantList passes a copy to PlantDetails. Not working
Removed the complicated mechanism to switch the location detail b/n edit/view modes in PlantDetails by adding it to the details table. Passing a copy to PlantDetails so that the actual values can be validated instead of the details strings.
BUG! Storing location name instead of location id in PlantDetails. Wanted to check in what I have so far anyway.